backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
python-multipart==0.0.12
Pillow==10.4.0
requests==2.32.3
